Added wolf
Added the first ennemy : a wolf.
Currently does no damage
It works on a state machine with 3 states : walking, running and attacking. The transition depend on the distance with the player. Each have their movement associated and animations.

TO FIX : Script in the sate machine could be cleaner, especially with the player and wolf speed. Current idea is to use a node (player, or state machine) to use to store those variable and make each state refer to this node.