Collision layer/mask + wolf attack
Changed collision IMPORTANT :
Collision layer is node toward others.
Collision Mask is others toward node
- Environment collisions are on layer 1 and 2.  Every entities will have mask 1. This will be used to prevent entities (player, enemies...) to go through StaticBody that has a collision (house, trees...).
- Enemies will have layer 2;  it can interact with StaticBody on layer 2, but cannot interact with the player.
- Player is layer 1 : Interact with both StaticBody and enemies.
- Layer 3 is used to define area (such as fire area). Only the player have mask 3, which make him the only one who can interact with these area.

Another change is the wolf attacking state. It now have a timer which start when entering the state. At the end of the timer, the distance to player is checked, and if it's below a certain range, then the player take damage (or will take damage, HP is not yet implemented). It also restart the timer. If the range is larger, it change the state back to running.