Added rabbit + Entities drops food
Added rabbit, which behavior is only to flee from the player if it get to close.
Both rabbit and wolf now have health and can die if the health get to 0. If it dies, the speed is reset to 0 and the sprite is replaced with raw meat. If the player get close to the raw meat, the player gain 15 hungriness and the entity's scene is finally queue.free()ed